move getEventsSummary to a helper funtion
