+++ apps/wallet/src/ui/app/helpers/index.ts
@@ -4,3 +4,4 @@
 export { default as formatDate } from './formatDate';
 export { default as notEmpty } from './notEmptyCheck';
 export { parseAmount } from './parseAmount';
+export { getEventsSummary } from './getEventsSummary';
